Move plot_interactive() to Map base class and make it work for hpx
