+++ gammapy/maps/base.py
@@ -934,6 +934,71 @@ class Map(object):
         """
         pass
 
+    def plot_interactive(self, axis=None, **kwargs):
+        """
+        Plot map with interactive widgets to explore the non spatial axes.
+
+        Parameters
+        ----------
+        axis : str
+            Axis name to slide through, with the interactive slider. By default
+            the first non-spatial axis is used.
+        **kwargs : dict
+            Keyword arguments passed to `WcsND.plot()`.
+
+        Examples
+        --------
+
+        You can try this out e.g. using a Fermi-LAT diffuse model cube with an energy axis::
+
+            %matplotlib inline
+            from gammapy.maps import Map
+
+            m = Map.read("$GAMMAPY_EXTRA/datasets/vela_region/gll_iem_v05_rev1_cutout.fits")
+            m.plot_interactive(cmap='gnuplot2')
+
+        If you would like to adjust the figure size you can use the `matplotlib.rc_context()`
+        context manager.
+
+            import matplotlib as mpl
+            with mpl.rc_context(rc={'figure.figsize': (12, 6)}):
+                m.plot_interactive()
+
+        """
+        from ipywidgets.widgets.interaction import interact, fixed
+        from ipywidgets import IntSlider, RadioButtons
+        import matplotlib.pyplot as plt
+
+        kwargs.setdefault('interpolation', 'nearest')
+        kwargs.setdefault('origin', 'lower')
+        kwargs.setdefault('cmap', 'afmhot')
+
+        if self.geom.is_image:
+            raise TypeError('Use .plot() for 2D Maps')
+
+        axis = axis or self.geom.axes[0].name
+        map_axis = self.geom.get_axis_by_name(axis)
+
+        @interact(
+            idx=IntSlider(min=0, max=map_axis.nbin - 1, step=1, value=0,
+                                    description=axis + ' idx'),
+            stretch=RadioButtons(options=['linear', 'sqrt', 'log'], value='sqrt',
+                                         description='Plot stretch'),
+        )
+        def _plot_interactive(idx, stretch):
+            img = self.get_image_by_idx([idx])
+            fig, ax, cbar = img.plot(stretch=stretch, **kwargs)
+
+            if map_axis.node_type == 'edge':
+                edges = map_axis.edges
+                title = '{:.0f} - {:.0f} '.format(edges[idx], edges[idx + 1])
+            else:
+                center = map_axis.center
+                title = '{:.0f} '.format(center[idx])
+            title += str(map_axis.unit)
+            ax.set_title(title)
+            plt.show()
+
     def copy(self, **kwargs):
         """Copy map instance and overwrite given attributes, except for geometry.
